Allineamento tramite casting a tipo long per quando riguarda id definiti come int

+++ Java/quadrangolare_calcio/src/main/java/com/example/quadrangolare_calcio/controller/AdminAllenatoreController.java
@@ -100,7 +100,7 @@ public class AdminAllenatoreController {
         if (session.getAttribute("admin") == null)
             return "redirect:/loginadmin";
 
-        Allenatore esistente = allenatoreService.dettaglioAllenatore(allenatore.getIdAllenatore());
+        Allenatore esistente = allenatoreService.dettaglioAllenatore((long) allenatore.getIdAllenatore());
 
         // NOME
         if (allenatore.getNome() != null && !allenatore.getNome().isBlank()) {
